Set autogenned methods' parents correctly
The ApiNode needs the accessibility info later to ensure we filter out private sets. It'd be nice to collapse them to {get; private set;} for example but I seem to recall there are edge cases where in IL the property names can be different from the normal auto-generated ones. It seems incredibly unlikely on a public API, so will probably opt for the nicer layout.